+++ src/sentry/disabled/disabled_sdk.h
@@ -17,6 +17,8 @@ class DisabledSDK : public InternalSDK {
 	virtual void set_user(const Ref<SentryUser> &p_user) override {}
 	virtual void remove_user() override {}
 
+	virtual void set_environment(const String &p_environment) override {}
+
 	virtual Ref<SentryBreadcrumb> create_breadcrumb() override { return memnew(DisabledBreadcrumb); }
 	virtual void add_breadcrumb(const Ref<SentryBreadcrumb> &p_breadcrumb) override {}